removed extraneous (double) cast per PR review

+++ src/modules/pipeline/worker.c
@@ -246,7 +246,7 @@ void ecs_workers_progress(
     }
 
     if (world->measure_frame_time) {
-        world->stats.system_time_total += (double)ecs_time_measure(&start);
+        world->stats.system_time_total += ecs_time_measure(&start);
     }    
 }
